RUE-532: constant u64 indices above i64::MAX get the compile-time bounds check (spec 7.1:9)
try_get_const_index_checked narrowed the evaluator's i128 through
ConstValue::as_integer() -> Option<i64>, so a u64 constant above i64::MAX
returned None — an exactly-known out-of-bounds index was indistinguishable
from a runtime index and deferred to the runtime trap (verified: exit 101
instead of E0902). The query now returns the full i128, the four bounds
checks compare in i128, and IndexOutOfBounds carries i128 so the diagnostic
reports the exact index (18446744073709551615, not a clamp).

Tests: 7.1:9 spec cases — huge-constant rejection pinning the exact index in
the message, and an in-bounds u64-constant control.

+++ crates/rue-error/src/lib.rs
@@ -1421,8 +1421,11 @@ pub enum ErrorKind {
     IndexOnNonArray { found: String },
     #[error("{}", format_array_length_mismatch(*.expected, *.found))]
     ArrayLengthMismatch { expected: u64, found: u64 },
+    // `index` is i128 so a constant u64 index above i64::MAX is still
+    // reported exactly (RUE-532) — the old i64 narrowing made such an index
+    // look non-constant and skip the compile-time bounds check entirely.
     #[error("index out of bounds: the length is {length} but the index is {index}")]
-    IndexOutOfBounds { index: i64, length: u64 },
+    IndexOutOfBounds { index: i128, length: u64 },
     #[error("type annotation required for empty array")]
     TypeAnnotationRequired,
     /// Cannot move non-Copy element out of array index position.
